Part 15 Cleaning Up Program Class (Refactor program.cs)
