+++ Talabat.APIs/Extensions/SwaggerServicesExtension.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.AspNetCore.Builder;
+
+namespace Talabat.APIs.Extensions
+{
+    public static class SwaggerServicesExtension
+    {
+        public static IServiceCollection  SwaggerServices(this IServiceCollection services) 
+        {
+            services.AddEndpointsApiExplorer();
+            services.AddSwaggerGen();
+
+            return services;
+        }
+
+        public static WebApplication UseSwaggerMiddleware(this WebApplication app)
+        {
+            app.UseSwagger();
+            app.UseSwaggerUI();
+
+            return app;
+        }
+    }
+}
